requirement.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
beautifulsoup4==4.12.2
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.1.5
djangorestframework==3.14.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
idna==3.4
kombu==5.3.1
lxml==4.9.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
numpy==1.25.1
pandas==2.0.3
prompt-toolkit==3.0.39
pyarrow==12.0.1
pylance==0.5.8
python-dateutil==2.8.2
pytz==2022.7
redis==4.6.0
redis-cli==0.1.2
redis-server==6.0.9
requests==2.31.0
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.4.1
sqlparse==0.4.3
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
